rename: 엔티티 및 레포지토리 패키지 domains 패키지로 내부로 이동
- 도메인 모듈에 여러 패키지 추가 필요에 따라 domains패키지 추가
