+++ domain/src/main/java/com/aireview/review/domains/category/Category.java
@@ -1,7 +1,7 @@
-package com.aireview.review.domain.category;
+package com.aireview.review.domains.category;
 
 
-import com.aireview.review.domain.BaseTimeEntity;
+import com.aireview.review.domains.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
