+++ domain/src/main/java/com/aireview/review/domains/category/CategoryRepository.java
@@ -1,4 +1,4 @@
-package com.aireview.review.domain.category;
+package com.aireview.review.domains.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
