+++ domain/src/main/java/com/aireview/review/domains/coupon/CouponRepository.java
@@ -1,4 +1,4 @@
-package com.aireview.review.domain.coupon;
+package com.aireview.review.domains.coupon;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
