+++ domain/src/main/java/com/aireview/review/domains/user/domain/UserRepository.java
@@ -1,4 +1,4 @@
-package com.aireview.review.domain.user;
+package com.aireview.review.domains.user.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,7 +8,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByEmail(Email email);
+    Optional<User> findByEmail(String email);
 
     Optional<User> findByOauthProviderAndOauthUserId(OAuthProvider oauthProvider, String oauthUserId);
 }
